requirements.txt
opencv-python~=4.10.0.84
numpy~=1.23.4
tensorflow~=2.17.0
tqdm~=4.66.1
Flask~=2.3.1
PyYAML~=6.0